moved data logging methods into their own class DataLogger as well as Error signal changing ones to ErrorStatus
