+++ src/managers/__init__.py
@@ -2,3 +2,5 @@ from .data_holder import DataHolder
 from .timer_service import TimerService
 from .device_manager import DeviceManager
 from .plot_manager import PlotManager
+from .data_logger import DataLogger
+from .error_status import ErrorStatus
